\name{CCs}
\docType{data}
\alias{contchoices}
\alias{CCs2}
\alias{CCs3}
\alias{CCs4}
\alias{CCs5}
\alias{CCs6}
\alias{CCs7}
\alias{CCs8}
\title{Continuous choices}
\description{
These data sets give the solutions found from the continuous optimization algorithm. Data sets CCs2, CCs3, CCs4, CCs5, CCs6, CCs7, and CCs8 give the choices in 2, 3, 4, 5, 6, 7, and 8 dimensions, respectively.
}
\usage{
data(CCs2);
data(CCs3);
data(CCs4);
data(CCs5);
data(CCs6);
data(CCs7);
data(CCs8);
}
\format{Matrices containing the choices.}
\references{
He, Xu (2024). "Efficient Kriging using interleaved lattice-based designs with low fill and high separation distance properties", \emph{SIAM/ASA Journal on Uncertainty Quantification}, 12(4): 1113-1134. 
}
