\name{GMs}
\docType{data}
\alias{GMs}
\alias{GMs2}
\alias{GMs3}
\alias{GMs4}
\alias{GMs5}
\alias{GMs6}
\alias{GMs7}
\alias{GMs8}
\title{Generator matrices of standard interleaved lattices, treating dimension permuted lattices as the same lattice}
\description{
These data sets give the generator matrices of standard interleaved lattices, treating dimension permuted lattices as the same lattice. Data sets GMs2, GMs3, GMs4, GMs5, GMs6, GMs7, and GMs8 give the matrices in 2, 3, 4, 5, 6, 7, and 8 dimensions, respectively.
}
\usage{
data(GMs2);
data(GMs3);
data(GMs4);
data(GMs5);
data(GMs6);
data(GMs7);
data(GMs8);
}
\format{Matrices containing generator matrices.}
\references{
He, Xu (2017). "Interleaved lattice-based minimax distance designs", \emph{Biometrika}, 104(3): 713-725.
}
