\name{GeneratorMatrices}
\docType{data}
\alias{GeneratorMatrices2}
\alias{GeneratorMatrices3}
\alias{GeneratorMatrices4}
\alias{GeneratorMatrices5}
\title{Generator matrices of standard interleaved lattices}
\description{
These data sets give the generator matrices of standard interleaved lattices, treating dimension prumeted lattices as different lattices. Data sets GeneratorMatrices2, GeneratorMatrices3, GeneratorMatrices4, and GeneratorMatrices5 give the matrices in 2, 3, 4, and 5, dimensions, respectively.
}
\usage{
data(GeneratorMatrices2);
data(GeneratorMatrices3);
data(GeneratorMatrices4);
data(GeneratorMatrices5);
}
\format{Matrices containing generator matrices.}
\references{
He, Xu (2018). "Interleaved lattice-based maximin distance designs", \emph{Biometrika}, accepted, arXiv:1807.02289v1. 
}
